+++ src/hap.ts
@@ -1,0 +1,6 @@
+import { HAP } from 'homebridge';
+
+export let hap: HAP;
+export function setHap(hapInstance: HAP) {
+  hap = hapInstance;
+}