+++ CHANGELOG.md
@@ -6,6 +6,10 @@ As soon as the project reaches a mature and stable state, the first major versio
 and after the project will apply [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][]
+### Added
+
+- Support for switches (relays) with more than two outputs (added a lot of keys starting with `state_`).
+
 ### Changed
 
 - Call to `registerPlatform` changed to include package name (see [#8](https://github.com/itavero/homebridge-z2m/issues/6)).