+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and after the project will apply [Semantic Versioning](https://semver.org/spec/v
 ### Changed
 
 - Call to `registerPlatform` changed to include package name (see [#8](https://github.com/itavero/homebridge-z2m/issues/6)).
+- Use global `hap` variable to access characteristics, instead of storing references in each ServiceWrapper.
 
 ## [0.0.5][] - 2020-07-05
 ### Changed