+++ CHANGELOG.md
@@ -6,6 +6,8 @@ As soon as the project reaches a mature and stable state, the first major versio
 and after the project will apply [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][]
+
+## [0.0.6][] - 2020-09-09
 ### Added
 
 - Support for switches (relays) with more than two outputs (added a lot of keys starting with `state_`).
@@ -49,7 +51,8 @@ and after the project will apply [Semantic Versioning](https://semver.org/spec/v
 - Improve state determination for WindowCovering.
 
 
-[Unreleased]: https://github.com/itavero/homebridge-z2m/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/itavero/homebridge-z2m/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/itavero/homebridge-z2m/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/itavero/homebridge-z2m/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/itavero/homebridge-z2m/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/itavero/homebridge-z2m/compare/v0.0.2...v0.0.3